Support GHC 8.10 and 8.8
Tempalte haskell bundled with GHC 8.10 has a breaking change in definition of
TupE, this is handled with CPP templating.

Template haskell bundled with GHC 8.10 has changed the behaviour for encoding
tuples with 1 item. Now they get encoded as 'GHC.Tuple.Unit' while earlier they
just got encoded as the type of the item in the tuple, this is handled by
checking for arity of the effect constructor as using the Unit type in tests is
almost never useful.